refactor devtools
